+++ CustomRoles/Configs/Roles/MedicZombieConfig.cs
@@ -8,7 +8,7 @@ namespace CustomRoles.Configs.Roles
         public RoleType RoleType { get; set; } = RoleType.Scp0492;
         public int AbilityCooldown { get; set; } = 90;
         public float AbilityDuration { get; set; } = 15f;
-        public float AbilityHealAmount { get; set; } = 10f;
-        public float AbilityFinaleProtection { get; set; } = 45f;
+        public float AbilityHealAmount { get; set; } = 10;
+        public ushort AbilityFinaleProtection { get; set; } = 45;
     }
 }
